added tab navigation in dashboard page, putting justification component and overview component inside tabs
